Add types.py, continue automaton impl
