+++ scapy/contrib/opcua/binary/automaton.py
@@ -1,60 +1,20 @@
 # coding=utf-8
-
-from scapy.all import Automaton, ATMT
-from scapy.contrib.opcua.binary.tcp import UaTcpHelloMessage, UaTcp
 import socket
 
+from scapy.automaton import Automaton
+import scapy.contrib.opcua.binary.types as UA
 
-class UaClient(Automaton):
-    """
-    This Automaton implements basic client functionality
-    """
+
+class _UaAutomaton(Automaton):
+
     def __init__(self, *args, **kargs):
-        super(UaClient, self).__init__(*args, **kargs)
-        self.socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        self.target = "localhost"
-        self.targetPort = 4840
+        super(_UaAutomaton, self).__init__(*args, **kargs)
+        self.socket = None
 
-    @ATMT.state(initial=1)
-    def START(self):
+    def parse_args(self, debug=0, store=1, **kwargs):
+        super(_UaAutomaton, self).parse_args(debug, store, **kwargs)
+
+    def recv(self):
+        len(UA.UaTcpMessageHeader())
+        self.socket.recv()
         pass
-
-    @ATMT.state()
-    def TCP_CONNECTED(self):
-        pass
-
-    @ATMT.state()
-    def CONNECTED(self):
-        pass
-
-    @ATMT.state()
-    def TCP_DISCONNECTED(self):
-        pass
-
-    @ATMT.state(final=1)
-    def END(self):
-        pass
-
-    @ATMT.condition(START)
-    def connectTCP(self):
-        self.socket.connect((self.target, self.targetPort))
-        raise self.TCP_CONNECTED()
-
-    @ATMT.condition(TCP_CONNECTED)
-    def connect(self):
-        self.send(UaTcp(Message=UaTcpHelloMessage()))
-        print("attempting to connect")
-        raise self.CONNECTED()
-
-    @ATMT.condition(CONNECTED)
-    def disconnect(self):
-        self.socket.close()
-        raise self.TCP_DISCONNECTED()
-
-    @ATMT.condition(TCP_DISCONNECTED)
-    def end(self):
-        raise self.END()
-
-    def my_send(self, pkt):
-        print(pkt)
-        self.socket.send(pkt.build())
