+++ scapy/contrib/opcua/binary/types.py
@@ -1,0 +1,9 @@
+# coding=utf-8
+"""
+This file gathers all default OPC UA types including networking.
+"""
+
+from scapy.contrib.opcua.binary.builtinTypes import *
+from scapy.contrib.opcua.binary.schemaTypes import *
+from scapy.contrib.opcua.binary.tcp import *
+from scapy.contrib.opcua.binary.secureConversation import *
